+++ src/CustomServerSettingsRequesterWindow.c
@@ -7,6 +7,7 @@
 #include "CustomServerSettingsRequesterWindow.h"
 #include "gui.h"
 
+Object *customServerTemplateCycle;
 Object *customServerHostString;
 Object *customServerPortString;
 Object *customServerUsesSSLCycle;
@@ -16,6 +17,51 @@ Object *customServerApiEndpointCycle;
 Object *customServerApiEndpointUrlString;
 Object *customServerFullUrlPreviewString;
 Object *customServerSettingsRequesterWindowObject;
+
+/* Template definitions */
+enum {
+    TEMPLATE_NONE = 0,
+    TEMPLATE_GOOGLE_GEMINI,
+    TEMPLATE_LM_STUDIO
+};
+
+HOOKPROTONHNONP(TemplateSelectedFunc, void) {
+    LONG template;
+    get(customServerTemplateCycle, MUIA_Cycle_Active, &template);
+
+    switch (template) {
+    case TEMPLATE_GOOGLE_GEMINI:
+        set(customServerHostString, MUIA_String_Contents,
+            "generativelanguage.googleapis.com");
+        set(customServerPortString, MUIA_String_Integer, 443);
+        set(customServerUsesSSLCycle, MUIA_Cycle_Active, 1); /* SSL enabled */
+        set(customServerChatModelString, MUIA_String_Contents,
+            "gemini-2.0-flash");
+        set(customServerApiEndpointCycle, MUIA_Cycle_Active,
+            API_ENDPOINT_CHAT_COMPLETIONS);
+        set(customServerApiEndpointUrlString, MUIA_String_Contents,
+            "v1beta/openai");
+        break;
+    case TEMPLATE_LM_STUDIO:
+        set(customServerHostString, MUIA_String_Contents, "localhost");
+        set(customServerPortString, MUIA_String_Integer, 1234);
+        set(customServerUsesSSLCycle, MUIA_Cycle_Active, 0); /* No SSL */
+        set(customServerChatModelString, MUIA_String_Contents, "");
+        set(customServerApiEndpointCycle, MUIA_Cycle_Active,
+            API_ENDPOINT_RESPONSES);
+        set(customServerApiEndpointUrlString, MUIA_String_Contents, "v1");
+        break;
+    default:
+        /* TEMPLATE_NONE - do nothing */
+        break;
+    }
+
+    /* Reset template cycle back to "None" after applying */
+    if (template != TEMPLATE_NONE) {
+        set(customServerTemplateCycle, MUIA_Cycle_Active, TEMPLATE_NONE);
+    }
+}
+MakeHook(TemplateSelectedHook, TemplateSelectedFunc);
 
 HOOKPROTONHNONP(SettingsChangedFunc, void) {
     STRPTR customServerHost;
@@ -92,11 +138,16 @@ MakeHook(CustomServerSettingsRequesterOkButtonClickedHook,
  * @return RETURN_OK on success, RETURN_ERROR on failure
  **/
 LONG createCustomServerSettingsRequesterWindow() {
+    static STRPTR templateOptions[4] = {NULL};
+    templateOptions[TEMPLATE_NONE] = STRING_MENU_TEMPLATE_NONE;
+    templateOptions[TEMPLATE_GOOGLE_GEMINI] = STRING_MENU_TEMPLATE_GOOGLE_GEMINI;
+    templateOptions[TEMPLATE_LM_STUDIO] = STRING_MENU_TEMPLATE_LM_STUDIO;
+
     static STRPTR sslOptions[3] = {NULL};
     sslOptions[0] = STRING_ENCRYPTION_NONE;
     sslOptions[1] = STRING_ENCRYPTION_SSL;
 
-    static STRPTR apiEndpointOptions[2] = {NULL};
+    static STRPTR apiEndpointOptions[3] = {NULL};
     apiEndpointOptions[0] = API_ENDPOINT_NAMES[API_ENDPOINT_RESPONSES];
     apiEndpointOptions[1] = API_ENDPOINT_NAMES[API_ENDPOINT_CHAT_COMPLETIONS];
 
@@ -108,6 +159,15 @@ LONG createCustomServerSettingsRequesterWindow() {
         MUIA_Window_Height, 200,
         MUIA_Window_CloseGadget, FALSE,
         WindowContents, VGroup,
+            Child, VGroup,
+                MUIA_Frame, MUIV_Frame_Group,
+                MUIA_FrameTitle, STRING_MENU_TEMPLATE,
+                Child, customServerTemplateCycle = CycleObject,
+                    MUIA_CycleChain, TRUE,
+                    MUIA_Cycle_Entries, templateOptions,
+                    MUIA_Cycle_Active, TEMPLATE_NONE,
+                End,
+            End,
             Child, VGroup,
                 MUIA_Frame, MUIV_Frame_Group,
                 MUIA_FrameTitle, STRING_PROXY_HOST,
@@ -205,6 +265,7 @@ LONG createCustomServerSettingsRequesterWindow() {
              MUIA_Pressed, FALSE, MUIV_Notify_Window, 3, MUIM_Set,
              MUIA_Window_Open, FALSE);
     DoMethod(customServerSettingsRequesterWindowObject, MUIM_Notify, MUIA_Window_Open, MUIV_EveryTime, MUIV_Notify_Application, 2, MUIM_CallHook, &SettingsChangedHook);
+    DoMethod(customServerTemplateCycle, MUIM_Notify, MUIA_Cycle_Active, MUIV_EveryTime, MUIV_Notify_Self, 2, MUIM_CallHook, &TemplateSelectedHook);
     DoMethod(customServerHostString, MUIM_Notify, MUIA_String_Contents, MUIV_EveryTime, MUIV_Notify_Self, 2, MUIM_CallHook, &SettingsChangedHook);
     DoMethod(customServerPortString, MUIM_Notify, MUIA_String_Contents, MUIV_EveryTime, MUIV_Notify_Self, 2, MUIM_CallHook, &SettingsChangedHook);
     DoMethod(customServerUsesSSLCycle, MUIM_Notify, MUIA_Cycle_Active, MUIV_EveryTime, MUIV_Notify_Self, 2, MUIM_CallHook, &SettingsChangedHook);